fixed repeat perms, added hmm check

+++ bot.js
@@ -7,17 +7,30 @@ bot.on("ready", () => {
 });
 
 bot.on("message", msg => {
-    if(msg.content.match(/good morning/i)){
+    //checks if good morning is said
+    if (msg.content.match(/good morning/i)) {
         msg.reply("It's afternoon...");
     }
+
+    //checks if ro says hmm
+    if (msg.content.match(/hmm/i)&&msg.member.id == config.roID) {
+        msg.channel.send("Correct.");
+    }
+
+    //idk what this does
     if (!msg.content.startsWith(config.prefix) || msg.author.bot) return;
 
+    //my first command <3
     if (msg.content.startsWith(config.prefix + "ping")) {
         msg.reply("POOOOOOOONG!");
     }
+    
+    //my second command
     if (msg.content.startsWith(config.prefix + "foo")) {
         msg.channel.send("bar!");
     }
+
+    //hugs 
     if (msg.content.startsWith(config.prefix + "hug")) {
         if (!msg.mentions.users.array()[0]) {
             msg.channel.send(`*hugs* ***${msg.author}***`);
@@ -27,12 +40,28 @@ bot.on("message", msg => {
             msg.channel.send(`*hugs* ***${huggee.username}***`);
         }
     }
+
+    //evil bot control for me and al
     if (msg.content.startsWith(config.prefix + "repeat")) {
-        let mess = msg.content;
-        msg.channel.send(mess.substring(8));
-        msg.delete();
+        if (msg.member.id == config.alID||msg.member.id == config.myUserID) {
+            let mess = msg.content;
+            msg.channel.send(mess.substring(8));
+            msg.delete();
+        } else {
+            msg.channel.send("hecc you, you're not allowed");
+        }
+    }
+    
+    //TODO fix
+    if (msg.content.startsWith(config.prefix + "repeatspam")) {
+        if (msg.member.id == config.myUserID) {
+            let mess = msg.content;
+            bot.channels.get(config.testID).send(mess.substring(12));
+            msg.delete();
+        } else {
+            msg.channel.send("hecc you, you're not allowed");
+        }
     }
 });
 
 bot.login(config.token);
-
